tests: separate some pgsql/mysql specifics
