+++ tests/pgsql-schema.sql
@@ -21,52 +21,23 @@ ALTER TABLE ak
 CREATE TYPE some_enum AS ENUM ('small','medium','large');
 CREATE TABLE typemix (
     pid SERIAL,
-    c_char5 varchar(5) NOT NULL,
-    c_varchar255 varchar(255) NOT NULL,
+    c_varchar5 varchar(5) NOT NULL,
     c_varchar25 varchar(25) DEFAULT NULL,
-    c_varbinary255 varchar NOT NULL,
-    c_varbinary25 varchar DEFAULT NULL,
+    c_varchar255 varchar NOT NULL,
     c_date date DEFAULT NULL,
     c_time time DEFAULT NULL,
     c_datetime timestamp DEFAULT NULL,
     c_timestamp timestamp NULL DEFAULT CURRENT_TIMESTAMP,
-    c_year smallint DEFAULT NULL,
-    c_tiny_text varchar DEFAULT NULL,
-    c_medium_text varchar DEFAULT NULL,
     c_text text,
-    c_long_text text,
     c_enum some_enum NOT NULL,
-    c_set varchar NOT NULL,
-
-
---     CBIT not tested now
---     c_bit255 bit(255) NOT NULL,
---     c_bit25 bit(25) DEFAULT NULL,
+    c_bit255 bit(255) NOT NULL,
+    c_bit25 bit(25) DEFAULT NULL,
     c_bit bit(7) DEFAULT NULL,
--- integer types included in tests however some types do not exist in psql
     c_int int NOT NULL,
-    c_tinyint smallint NOT NULL,
     c_smallint smallint NOT NULL,
-    c_mediumint smallint NOT NULL,
     c_bigint bigint NOT NULL,
-    c_double float NOT NULL,
-    c_real float NOT NULL,
-    c_boolean boolean NOT NULL,
-
---     NOT TESTED
---     c_bytea bytea NOT NULL,
-
--- TESTED but types do not exist in psql
-    c_blob varchar NOT NULL,
-    c_tinyblob varchar NOT NULL,
-    c_mediumblog varchar NOT NULL,
-    c_longblob varchar NOT NULL,
-    c_unsigned_tinyint smallint NOT NULL,
-    c_unsigned_int int NOT NULL,
-    c_unsigned_smallint smallint NOT NULL,
-    c_unsigned_mediumint smallint NOT NULL,
-    c_unsigned_bigint bigint NOT NULL
---
+    c_float float NOT NULL,
+    c_boolean boolean NOT NULL
 );
 
 ALTER TABLE typemix
